Refactor authentication in form.jsx
