+++ src/forms/formValidation.js
@@ -1,5 +1,7 @@
-// eslint-disable-next-line import/prefer-default-export
-export function validateInput(input, errMessage) {
+import { Observable } from 'rxjs/Observable';
+import 'rxjs/add/operator/map';
+
+function validateInput(input, errMessage) {
   const errorLabel = input.parentNode.querySelector('.error-feedback');
 
   if (input.validity.valid) {
@@ -10,3 +12,17 @@ export function validateInput(input, errMessage) {
     errorLabel.textContent = errMessage;
   }
 }
+
+const authenticateFormData = data =>
+  Observable.from(data)
+    .filter(fields => fields.nodeName.toLowerCase() !== 'button')
+    .map(input => {
+      const { dataset, validationMessage } = input;
+      const message = dataset.validationmessage || validationMessage;
+
+      validateInput(input, message);
+
+      return input;
+    });
+
+export default authenticateFormData;
